+++ changelog.md
@@ -3,6 +3,7 @@
 All notable changes to this project will be documented in this file.
 
 ## 2021.09
+* Fixed: Style guide typography regression caused by Core's reset.css enqueuing in the block editor with the v5.8.0 release of WordPress.
 * Updated: ci GitHub action to allow manual runs
 * Fixed: Replace deleted repo https://github.com/hautelook/phpass with https://github.com/bordoni/phpass
 * Updated: wp-browser to 3.0.9